requirements.txt
# MainFrame Code Generator - Dependencies

# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0
rich>=13.7.0
pyyaml>=6.0.1

# Code Analysis & Validation
lxml>=5.1.0

# CLI & UI
click>=8.1.7
questionary>=2.0.1

# Utilities
python-dateutil>=2.8.2
jinja2>=3.1.3

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type Checking
mypy>=1.8.0
types-requests>=2.31.0

